refactor: Codex app-server クライアントを internal/infra/codexapp へ切り出し (layered 4/10)
cmd/fanout/codex_plan_tui.go(1,447 行)は WebSocket 手実装 + JSON-RPC +
app-server プロセス管理 + Plan Mode コントローラ + status ファイル握手の
自己完結インフラ実装だった。internal/infra/codexapp パッケージへ移し、
cmd 側は parse/dispatch のみ 106 行に縮小する。

- 公開 API: Client/Dial/Requester/SessionClient、Server/StartServer/
  Connect、RunPlanTUI(TUIConfig)、Status/WriteStatus/ReadStatus/WaitReady、
  LaunchCommand/ResumeLaunchCommand/StatusPath
- initialize clientInfo の version は TUIConfig.Version で注入し、
  cmd の ldflags 変数 version を渡す(挙動不変)
- エラー文言・status JSON タグ・生成コマンド文字列・exitcode は
  verbatim 移送(主要文言 53 種の新旧突き合わせで差分ゼロ)
- テスト移送: parse 系 2 本は cmd に残し、残り全部 + pane_test の
  WaitReady 系 3 本を codexapp の白箱テストへ
