+++ internal/infra/codexapp/launch.go
@@ -1,0 +1,92 @@
+package codexapp
+
+import (
+	"fmt"
+	"os"
+	"path/filepath"
+	"strings"
+	"time"
+
+	"github.com/butaosuinu/fanout/internal/core/agent"
+)
+
+// PlanTUICommand is the hidden fanout subcommand that runs the Plan Mode
+// controller inside the child pane.
+const PlanTUICommand = "__codex-plan-tui"
+
+// LaunchCommand builds the shell command that starts the Plan Mode controller
+// for a fresh prompt.
+func LaunchCommand(fanoutPath, codexPath, prompt, statusPath string) string {
+	if strings.TrimSpace(fanoutPath) == "" {
+		fanoutPath = "fanout"
+	}
+	args := []string{
+		fanoutPath,
+		"__codex-plan-tui",
+		"--codex", codexPath,
+		"--prompt", prompt,
+		"--status-file", statusPath,
+	}
+	quoted := make([]string, len(args))
+	for i, arg := range args {
+		quoted[i] = agent.ShellQuote(arg)
+	}
+	return strings.Join(quoted, " ")
+}
+
+// ResumeLaunchCommand builds the shell command that resumes an existing Plan
+// Mode thread in a restored pane.
+func ResumeLaunchCommand(fanoutPath, codexPath, threadID, sessionID, statusPath string) string {
+	if strings.TrimSpace(fanoutPath) == "" {
+		fanoutPath = "fanout"
+	}
+	args := []string{
+		fanoutPath,
+		"__codex-plan-tui",
+		"--codex", codexPath,
+		"--resume-thread-id", threadID,
+	}
+	if strings.TrimSpace(sessionID) != "" {
+		args = append(args, "--resume-session-id", sessionID)
+	}
+	args = append(args, "--status-file", statusPath)
+	quoted := make([]string, len(args))
+	for i, arg := range args {
+		quoted[i] = agent.ShellQuote(arg)
+	}
+	return strings.Join(quoted, " ")
+}
+
+// StatusPath derives the /tmp status file path for one Plan Mode launch. Dry
+// runs use a deterministic name so goldens stay stable.
+func StatusPath(projectRoot string, issueNum int, dryRun bool) string {
+	repo := safeCodexPlanTempPart(filepath.Base(projectRoot))
+	base := fmt.Sprintf("fanout-codex-plan-%s-%d", repo, issueNum)
+	if dryRun {
+		return filepath.Join("/tmp", base+".json")
+	}
+	unique := fmt.Sprintf("%s-%d-%d", base, os.Getpid(), time.Now().UnixNano())
+	return filepath.Join("/tmp", unique+".json")
+}
+
+func safeCodexPlanTempPart(s string) string {
+	var b strings.Builder
+	for _, r := range s {
+		switch {
+		case r >= 'a' && r <= 'z',
+			r >= 'A' && r <= 'Z',
+			r >= '0' && r <= '9',
+			r == '.', r == '-', r == '_':
+			b.WriteRune(r)
+		default:
+			b.WriteByte('-')
+		}
+		if b.Len() >= 40 {
+			break
+		}
+	}
+	if b.Len() == 0 {
+		return "repo"
+	}
+	return b.String()
+}
