+++ internal/infra/codexapp/status_test.go
@@ -1,0 +1,42 @@
+package codexapp
+
+import (
+	"path/filepath"
+	"strings"
+	"testing"
+	"time"
+)
+
+func TestWaitForCodexPlanTUIReadyReadsReadyStatus(t *testing.T) {
+	path := filepath.Join(t.TempDir(), "status.json")
+	if err := writeStatus(path, Status{Status: statusReady}); err != nil {
+		t.Fatal(err)
+	}
+
+	if err := waitForCodexPlanTUIReady(path, time.Second); err != nil {
+		t.Fatalf("waitForCodexPlanTUIReady() failed: %v", err)
+	}
+}
+
+func TestWaitForCodexPlanTUIReadyReturnsFailedStatus(t *testing.T) {
+	path := filepath.Join(t.TempDir(), "status.json")
+	if err := writeStatus(path, Status{Status: statusFailed, Error: "boom"}); err != nil {
+		t.Fatal(err)
+	}
+
+	err := waitForCodexPlanTUIReady(path, time.Second)
+
+	if err == nil || err.Error() != "boom" {
+		t.Fatalf("waitForCodexPlanTUIReady() error = %v, want boom", err)
+	}
+}
+
+func TestWaitForCodexPlanTUIReadyTimesOutWithoutStatus(t *testing.T) {
+	path := filepath.Join(t.TempDir(), "missing.json")
+
+	err := waitForCodexPlanTUIReady(path, time.Millisecond)
+
+	if err == nil || !strings.Contains(err.Error(), errCodexPlanStartupTimeout.Error()) {
+		t.Fatalf("waitForCodexPlanTUIReady() error = %v, want timeout", err)
+	}
+}
